chroma-agents/requirements.txt
chromadb==0.4.15
sentence-transformers==2.2.2
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3